pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "project"
authors = ["hamzaqasem"]
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11.2"
fastapi = "^0.92.0"
motor = "^3.1.1"
gunicorn = "^20.1.0"
PyYAML = "^6.0"
uvicorn = "^0.20.0"
python-dotenv = "^1.0.0"
psutil = "^5.9.4"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
websockets = "^12.0"
aioredis = "^2.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
aiologger = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.20.3"
mypy = "^1.1.1"
requests = "^2.28.2"

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"

# [tool.poetry.scripts]
# start = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
